grocery_search_app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
psycopg2-binary==2.9.9
supabase==2.0.2
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1
rapidfuzz==3.5.2
pydantic==2.5.0
httpx==0.25.2